# uint256: minor improvement for Mul