Minor tweek of the code. Not sure where the boost comes from. But it does improve around 10% in my two x86-64 machines.

godbolt links:
old:  https://go.godbolt.org/z/z8vqhG9r7
new: https://go.godbolt.org/z/M17Ycfzxx
## Test
```
go test ./...
```
```
ok  	github.com/holiman/uint256	1.380s
```
## Benchmark
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: AMD Ryzen 7 7735H with Radeon Graphics         
                      │     old     │                 new                 │
                      │   sec/op    │   sec/op     vs base                │
Mul/single/uint256-16   4.239n ± 2%   3.725n ± 1%  -12.14% (p=0.000 n=10)
Mul/single/big-16       37.05n ± 1%   36.96n ± 2%        ~ (p=0.403 n=10)
geomean                 12.53n        11.73n        -6.38%
```
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: Intel Core Processor (Broadwell, no TSX, IBRS)
                   │     old      │                new                 │
                   │    sec/op    │   sec/op     vs base               │
Mul/single/uint256   10.125n ± 4%   9.346n ± 2%  -7.69% (p=0.000 n=10)
Mul/single/big        79.24n ± 3%   79.11n ± 5%       ~ (p=0.853 n=10)
geomean               28.32n        27.19n       -4.00%
```