@holiman 2024-05-14
Interesting! 
```
name                  old time/op  new time/op  delta
Mul/single/uint256-8  15.1ns ± 0%  10.6ns ±35%  -29.68%  (p=0.000 n=8+10)
```
@chfast lgty?