@chfast 2024-05-14
> New: https://go.godbolt.org/z/5E8TahK9j Old: https://go.godbolt.org/z/nY31bsxdb
> 
> With Go 1.22.1, the new code has 7 less `MOVQ` instructions. Other instructions are the same amount. Possibly, this is the reason.

godbolt has also the "diff view": https://go.godbolt.org/z/ddcK6TzKK

It looks the new code spills less values from registers to stack. Good change.